Fix some initialization ordering issues
The user needs to set `mailaprop-address-file' themselves.  Don't even
try to initialize it in mailaprop.el, and throw an error in
`mailaprop-load-addresses' if it has not yet been set.

Check that `mailaprop-skip-address-fn' is non-nil before invoking it,
although that probably wasn't the cause of the initialization problems
that initially motivated this commit.