Don't initialize `mailaprop-addresses'
The documentation tells the user to initialize `mailaprop-addresses'
themselves.  Therefore, mailaprop.el should not do it at load time,
because if it were to do so, then `mailaprop-address-file' would need
to be set by the user before mailaprop.el were loaded, and it is not
good to have an unexpected ordering dependency like that.